+++ pkg/draw/polygon.go
@@ -36,7 +36,7 @@ func (p Polygon) Fill(d displayer, clr color.Color) {
 		for i := 0; i < len(p); i++ {
 			next := (i + 1) % len(p)
 			if intersects(p[i], p[next], y) {
-				x := intersectionX(p[i], p[next], y)
+				x := float64(intersectionX(p[i], p[next], y))
 				intersections = append(intersections, x)
 			}
 		}
@@ -46,8 +46,8 @@ func (p Polygon) Fill(d displayer, clr color.Color) {
 
 		// Draw horizontal lines between pairs of intersections.
 		for i := 0; i < len(intersections)-1; i += 2 {
-			start := geom.MakePoint(intersections[i], y)
-			end := geom.MakePoint(intersections[i+1], y)
+			start := geom.MakePoint(float32(intersections[i]), y)
+			end := geom.MakePoint(float32(intersections[i+1]), y)
 			line := Line(geom.MakeLine(start, end))
 			line.Draw(d, clr)
 		}
@@ -55,12 +55,12 @@ func (p Polygon) Fill(d displayer, clr color.Color) {
 }
 
 // Determines if a line segment intersects with a horizontal line at y.
-func intersects(p1, p2 geom.Point, y float64) bool {
+func intersects(p1, p2 geom.Point, y float32) bool {
 	return (p1.Y <= y && p2.Y > y) || (p1.Y > y && p2.Y <= y)
 }
 
 // Gets the x-coordinate of intersection of a line segment with a horizontal line at y.
-func intersectionX(p1, p2 geom.Point, y float64) float64 {
+func intersectionX(p1, p2 geom.Point, y float32) float32 {
 	if p1.Y == p2.Y {
 		return p1.X
 	}
